Suppress linting issues and improve type handling across modules:
- Replace `# pylint: disable=...` with `# noqa: B023` for usage with ruff linter.
- Adjust type coercion and defaults in `mask.py` to improve floating-point precision handling.
- Add explicit naming for unused iteration variables in `deep_typed_graph_net.py`.
- Refactor exception handling in `test_dask_mpi.py` to include chained exceptions.
- Remove unnecessary `cast` usage in `autoregressive.py`.
- Make minor updates to variable naming for readability in `plot_utils.py`.
- Simplify dataset reader and postprocessing logic in `dataset.py` while maintaining linting suppression.

+++ graphcast/mask.py
@@ -11,10 +11,10 @@ class Mask(predictor_base.Predictor):
     self,
     predictor: predictor_base.Predictor,
     mask: xr.DataArray,
-    value: np.floating | float = np.float32(0.0),
+    value: np.floating | float = 0.0,
   ):
     self._predictor = predictor
-    self._value = value
+    self._value = np.float32(value)
     self._mask = mask
 
   def __call__(
